Fix Input Actions, Rename & Improve StateDebugger.cs
Fixed the issues with the input system so that it now allows the user to play with a keyboard and controller, or two controllers.
Also made work on the StateDebugger (formally Debugger.cs) so that it no longer compiles outside the editor. (#if UNITY_EDITOR)
