+++ ProjectSettings/CommonBurstAotSettings.json
@@ -1,6 +1,6 @@
 {
   "MonoBehaviour": {
-    "Version": 3,
+    "Version": 4,
     "DisabledWarnings": ""
   }
 }
